+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 ## [Unreleased]
 ### Fixed
-- Errors are not reported 
+- Errors not being reported to crash tracker
 
 ## [1.3.4] - 2022-05-04
 ### Fixed