+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 ## [Unreleased]
 ### Added
 - Support for preview and prerelease plugins ([#38](https://github.com/pszklarska/FlutterPubVersionChecker/issues/38))
+### Fixed
+- Errors not being reported to crash tracker
 
 ## [1.3.4] - 2022-05-04
 ### Fixed